% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.deconvolution.R
\name{DeMixSC.deconvolution}
\alias{DeMixSC.deconvolution}
\title{Parallel execution of DeMixSC.core for deconvolution}
\usage{
DeMixSC.deconvolution(
  bulk.mat,
  ref,
  iter.max = 2000,
  eps = 0.001,
  nthread = (detectCores()[1] - 1),
  print.sample = 1,
  output.file = "./DeMixSC.save.iter.txt"
)
}
\arguments{
\item{bulk.mat}{A expression matrix, which genes on the rows and samples on the columns.}

\item{ref}{Reference matrix of expression signatures, with genes on the rows and cell types on the columns.}

\item{iter.max}{Maximum number of iterations for the algorithm. Default is 2000.}

\item{eps}{Convergence threshold. The iterations stop when the change between iterations is below this threshold.
Default is 0.001.}

\item{nthread}{Number of threads allocated for deconvolution. Default is using one less than total available threads.}

\item{print.sample}{Sample index for which iterative proportions should 
be recorded and written to the output file. This can be useful for 
monitoring the performance on a specific sample. Default is sample 1.}

\item{output.file}{Name of the output file where iterative data for 
the selected sample will be written.}
}
\value{
\item{proportion}{A matrix of cell-type proportions, with cell types on the rows and samples on the columns.}
 \item{converge}{A vector indicating the number of iterations until convergence for each sample.}
 \item{weights}{A matrix of gene weights with genes, weith genes on the rows and samples on the columns.}
}
\description{
Executes the DeMixSC.core function in parallel for estimating cell-type proportions.
}
\note{
By default, this function utilizes the maximum number of available cores minus one for parallel computation. 
This is designed to maximize computational efficiency. 
However, users should be cautious when running other tasks concurrently to avoid potential system slowdowns or overloading.
If resource concerns arise or if you're operating on a shared system, users might consider specifying a lower number of threads through the `nthread` parameter.
}
