% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.preprocess.R
\name{DeMixSC.discrepancy}
\alias{DeMixSC.discrepancy}
\title{Identify genes less affected by technological discrepancies}
\usage{
DeMixSC.discrepancy(mat.a, mat.b, cutoff = 0.05)
}
\arguments{
\item{mat.a}{Expression matrix 'a' from the benchmark dataset, with genes as rows and cell types as columns (potentially the bulk matrix).}

\item{mat.b}{Expression matrix 'b' from the benchmark dataset, with genes as rows and cell types as columns (potentially the pseudo-bulk matrix).}

\item{cutoff}{A cutoff P-value threshold to classify genes influenced by technological discrepancies. Default is 0.05.}
}
\value{
\item{non.discrepancies.genes}{A list of genes minimally influenced by technological discrepancies.}
 \item{mat.a.normalized}{Normalized expression matrix 'a' from the benchmark dataset.}
 \item{mat.b.normalized}{Normalized expression matrix 'b' from the benchmark dataset.}
 \item{t.test.res}{Paired t-test results table for genes detailing various parameters, with genes on the rows and parameters on the columns.}
}
\description{
Using the benchmark dataset as input, this function discerns genes that remain stable across technological platforms from those that exhibit high discrepancies.
}
