% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.preprocess.R
\name{DeMixSC.prep.benchmark}
\alias{DeMixSC.prep.benchmark}
\title{Preprocess a benchmark dataset for DeMixSC deconvolution.}
\usage{
DeMixSC.prep.benchmark(
  Seurat.obj,
  annotation,
  mat.a,
  mat.b,
  cutoff = 0.05,
  scale.factor = 1e+06
)
}
\arguments{
\item{Seurat.obj}{The pre-annotated single-cell Seurat object.}

\item{annotation}{A column name or scalar indicating cell-type information within the Seurat object.}

\item{mat.a}{mRNA expression matrix 'a' from the benchmark dataset (possibly bulk matrix), with genes on the rows and cell types on the columns.}

\item{mat.b}{mRNA expression matrix 'b' from the benchmark dataset (possibly pseudo-bulk matrix), with genes on the rows and cell types on the columns.}

\item{cutoff}{Threshold to differentiate genes based on inter-platform discrepancies. Default is 0.05.}

\item{scale.factor}{Scale factor applied to the relative expression abundances matrix. Default is 1e6.}
}
\value{
\item{mat.a.adj}{Adjusted relative expression matrix 'a', with genes on the rows and samples on the columns.}
 \item{mat.b}{Relative expression matrix 'b', with genes on the rows and samples on the columns.}
 \item{reference.adj}{Adjusted reference matrix used for deconvolving mat.a.adj}
 \item{reference}{Original reference matrix used for deconvolving mat.b}
 \item{theta.adj}{Adjusted relative abundance matrix of reference expression.}
 \item{cell.size}{A vector of the mean cell size for each cell type.}
 \item{discrepancy.genes}{A list of genes highly affected with technological discrepancies.}
 \item{non.discrepancy.genes}{A list of genes hardly affected by technological discrepancies.}
}
\description{
Prepares and adjusts the benchmark dataset to handle technological discrepancies for the DeMixSC deconvolution.
}
