% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.preprocess.R
\name{DeMixSC.prep}
\alias{DeMixSC.prep}
\title{Preprocess a benchmark dataset for deconvolution of unmatched bulk cohort using DeMixSC}
\usage{
DeMixSC.prep(
  Seurat.obj,
  annotation,
  mat.a,
  mat.b,
  mat.target,
  cutoff = 0.05,
  scale.factor = 1e+06,
  combat = TRUE
)
}
\arguments{
\item{Seurat.obj}{The pre-annotated single-cell Seurat object.}

\item{annotation}{A column name or scalar indicating cell-type information within the Seurat object.}

\item{mat.a}{mRNA expression matrix 'a' from the benchmark dataset (possibly bulk matrix), with genes on the rows and cell types on the columns.}

\item{mat.b}{mRNA expression matrix 'b' from the benchmark dataset (possibly pseudo-bulk matrix), with genes on the rows and cell types on the columns.}

\item{mat.target}{Targeted bulk cohort matrix for deconvolution, with genes on the rows and cell types on the columns.}

\item{cutoff}{Threshold to differentiate genes based on inter-platform discrepancies. Default is 0.05.}

\item{scale.factor}{Scale factor applied to the relative expression abundances matrix. Default is 1e6.}

\item{combat}{Logical indicator for using ComBat to align the unmatched bulk cohort with the benchmark dataset's bulk data. Default is TRUE.}
}
\value{
\item{mat.target.adj}{Adjusted target bulk cohort's relative expression matrix.}
 \item{reference.adj}{Adjusted reference matrix suitable for deconvolution}
 \item{theta.adj}{Adjusted relative abundance matrix of reference expression.}
 \item{cell.size}{A vector of the mean cell size for each cell type.}
 \item{discrepancy.genes}{A list of genes highly affected with technological discrepancies.}
 \item{non.discrepancy.genes}{A list of genes hardly affected by technological discrepancies.}
}
\description{
Adjusts and prepares the target bulk matrix and a reference matrix for deconvolution.
}
