% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.preprocess.R
\name{DeMixSC.preprocessing}
\alias{DeMixSC.preprocessing}
\title{Preprocess the input data for DeMixSC wNNLS framework}
\usage{
DeMixSC.preprocessing(
  mat.target = NULL,
  mat.a = NULL,
  mat.b = NULL,
  reference = NULL,
  min.expression = 1,
  scale.factor = 1e+05,
  adjp.cutoff = 0.05,
  log2fc.cutoff = 0.75,
  top.ranked.genes = 5000,
  adj.factor = 1000
)
}
\arguments{
\item{mat.target}{A gene expression matrix of target bulk cohort data to be deconvolved, where rows are genes and columns are samples.
If `NULL`, the function runs in benchmark mode (i.e., deconvolves the benchmark data `mat.a` and `mat.b`).}

\item{mat.a}{A gene expression matrix 'a' from the benchmark dataset (e.g., bulk matrix), where rows are genes and columns are samples.}

\item{mat.b}{A gene expression matrix 'b' from the benchmark dataset (e.g., pseudo-bulk matrix), where rows are genes and columns are samples.}

\item{reference}{A cell-type-specific gene expression matrix, where rows are genes and columns are cell types.}

\item{min.expression}{A numeric value that defines the minimum expression threshold for filtering genes.
Genes with expression values below this threshold across all samples will be removed. Default is 1.}

\item{scale.factor}{A numeric value used to scale the gene expression data.
This helps normalize the data by making the total expression in each sample comparable. Default is 1e5.}

\item{adjp.cutoff}{A numeric value representing the adjusted p-value cutoff (a threshold for statistical significance after accounting for multiple tests).
Genes with adjusted p-values greater than this threshold are not considered significantly differentially expressed. Default is 0.05.}

\item{log2fc.cutoff}{A numeric value defining the log2 fold change cutoff for selecting differentially expressed genes.
Genes with a log2 fold change smaller than this threshold are not considered. Default is 0.75.}

\item{top.ranked.genes}{An integer value indicating the number of top-ranked differentially expressed genes to be selected based on their mean expression.
These genes are used for technological discrepancy adjustments. Default is 5000.}

\item{adj.factor}{A numeric value used as a scaling factor to adjust the expression levels of discrepancy genes. Default is 1000.}
}
\value{
**Benchmark Mode:**
When `mat.target` is not provided (i.e., benchmark mode), the function returns:
 \item{mat.a.adj}{Adjusted relative expression matrix 'a', with genes on the rows and samples on the columns.}
 \item{mat.b}{Relative expression matrix 'b', with genes on the rows and samples on the columns.}
 \item{reference.adj}{Adjusted reference matrix used for deconvolving mat.a.adj.}
 \item{reference}{Original reference matrix used for deconvolving mat.b.}
 \item{discrepancy.genes}{A list of genes highly affected with technological discrepancies.}
 \item{non.discrepancy.genes}{A list of genes without detected technological discrepancies.}
 \item{de.res}{Results of differential expression analysis between bulk and pseudobulk data, including p-values, fold changes, and mean expression levels.}

**Real Data Mode:**
When `mat.target` is provided, the function returns:
 \item{mat.target.adj}{Adjusted expression matrix for the target bulk cohort after technological discrepancy correction.}
 \item{reference.adj}{Adjusted reference matrix after correcting for technological discrepancies.}
 \item{discrepancy.genes}{A list of genes highly affected with technological discrepancies.}
 \item{non.discrepancy.genes}{A list of genes without detected technological discrepancies.}
 \item{de.res}{Results of differential expression analysis between bulk and pseudobulk data, including p-values, fold changes, and mean expression levels.}
}
\description{
This function preprocesses input data for DeMixSC by handling two modes:
1) real data mode when target bulk data is provided, and
2) benchmark mode when target bulk data is not provided.
The function performs filtering, normalization, and adjusts genes with technological discrepancies.
}
\author{
Shuai Guo, Xiaoqian Liu, Wenyi Wang
}
