% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.R
\name{DeMixSC}
\alias{DeMixSC}
\title{The main function of DeMixSC package}
\usage{
DeMixSC(
  option = c("retina", "hgsc", "user.defined"),
  benchmark.mode = TRUE,
  mat.target = NULL,
  mat.a = NULL,
  mat.b = NULL,
  reference = NULL,
  min.expression = 1,
  scale.factor = 1e+05,
  adjp.cutoff = 0.05,
  log2fc.cutoff = 0.75,
  top.ranked.genes = 5000,
  adj.factor = 1000,
  c = 2,
  iter.max = 2000,
  eps = 0.001,
  nthread = (detectCores()[1] - 1),
  print.sample = 1,
  output.file = "./DeMixSC.save.iter.txt"
)
}
\arguments{
\item{option}{A character string specifying the mode of operation.
Options include:
- `"retina"`: Runs the deconvolution using the provided retina benchmark data and consensus reference.
- `"hgsc"`: Runs the deconvolution using the provided high-grade serous ovarian carcinoma (HGSC) benchmark data and consensus reference.
- `"user.defined"`: Allows the user to provide their own reference, benchmark, and bulk cohort data for deconvolution.}

\item{benchmark.mode}{Indicates whether DeMixSC should run in benchmark mode.
When set to TRUE (the default), DeMixSC will use predefined benchmark datasets for deconvolution.
If set to FALSE, the user must provide their own target bulk matrix (`mat.target`) for deconvolution in real data mode.}

\item{mat.target}{A gene expression matrix of target bulk cohort data to be deconvolved, where rows are genes and columns are samples.
If `NULL`, the function runs in benchmark mode (i.e., deconvolves the benchmark data `mat.a` and `mat.b`).}

\item{mat.a}{A gene expression matrix 'a' from the benchmark dataset (e.g., bulk matrix), where rows are genes and columns are samples.
Used when `option = "user.defined"`.}

\item{mat.b}{A gene expression matrix 'b' from the benchmark dataset (e.g., pseudo-bulk matrix), where rows are genes and columns are samples.
Used when `option = "user.defined"`.}

\item{reference}{A cell-type-specific gene expression matrix, where rows are genes and columns are cell types.}

\item{min.expression}{A numeric value that defines the minimum expression threshold for filtering genes.
Genes with expression values below this threshold across all samples will be removed. Default is 1.}

\item{scale.factor}{A numeric value used to scale the gene expression data.
This helps normalize the data by making the total expression in each sample comparable. Default is 1e5.}

\item{adjp.cutoff}{A numeric value representing the adjusted p-value cutoff (a threshold for statistical significance after accounting for multiple tests).
Genes with adjusted p-values greater than this threshold are not considered significantly differentially expressed. Default is 0.05.}

\item{log2fc.cutoff}{A numeric value defining the log2 fold change cutoff for selecting differentially expressed genes.
Genes with a log2 fold change smaller than this threshold are not considered. Default is 0.75.}

\item{top.ranked.genes}{An integer value indicating the number of top-ranked differentially expressed genes to be selected based on their mean expression.
These genes are used for technological discrepancy adjustments. Default is 5000.}

\item{adj.factor}{A numeric value used as a scaling factor to adjust the expression levels of discrepancy genes. Default is 1000.}

\item{c}{A numeric value representing the tuning parameter for the weight function in the weighted non-negative least squares algorithm. Default is 2.}

\item{iter.max}{An integer specifying the maximum number of iterations for the deconvolution algorithm.
The algorithm will stop if this number is reached, even if convergence (steady results over iterations) has not been achieved. Default is 2000.}

\item{eps}{A numeric value representing the convergence threshold for the deconvolution algorithm.
The algorithm will stop if the difference between iterations falls below this threshold. Default is 0.001.}

\item{nthread}{An integer specifying the number of threads to use for parallel processing (executing multiple operations simultaneously).
By default, it uses one less than the total number of available CPU cores.}

\item{print.sample}{An integer specifying the index of the sample whose iterative results should be printed during the deconvolution process.
This is useful for monitoring the algorithm’s progress on a specific sample. Default is 1.
Users can set `print.sample = NULL` to disable saving iteration details.}

\item{output.file}{A character string specifying the name of the file where the algorithm’s iterative results will be saved.
This file stores intermediate deconvolution results for the specified sample. Default is `"./DeMixSC.save.iter.txt"`.}
}
\value{
**Benchmark Mode:**
When `mat.target` is not provided (i.e., benchmark mode), the function returns:
\item{bulk.prop}{A matrix of estimated cell-type proportions for the bulk benchmark data (`mat.a`), where rows are cell types and columns are samples.}
\item{pseudobulk.prop}{A matrix of estimated cell-type proportions for the pseudo-bulk benchmark data (`mat.b`), where rows are cell types and columns are samples.}

**Real Data Mode:**
When `mat.target` is provided, the function returns:
 \item{proportion}{A matrix of estimated cell-type proportions for the provided bulk cohort (`mat.target`), with rows are cell types and columns are samples.}
 \item{converge}{A vector indicating the number of iterations for the deconvolution algorithm to converge for each sample in the bulk cohort.}
}
\description{
Implements an integrated workflow for preprocessing and deconvolving bulk RNA-seq datasets using single-cell reference data.
}
\author{
Shuai Guo, Xiaoqian Liu, Wenyi Wang
}
