% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.reference.R
\name{DeMixSC.ref}
\alias{DeMixSC.ref}
\title{Build the reference matrix from given Seurat object}
\usage{
DeMixSC.ref(Seurat.obj, annotation)
}
\arguments{
\item{Seurat.obj}{The pre-annotated single-cell Seurat object.}

\item{annotation}{A scalar or a character string to designate the column for cell-type information in the Seurat object.}
}
\value{
\item{reference}{A matrix of the reference expression,
                 with genes on the rows and cell types on the columns.}
\item{theta}{A matrix of the expression abundance relative to total UMI count per cell type,
             with genes on the rows and cell types on the columns.}
\item{cell.size}{A vector of the mean cell size for each cell type.}
}
\description{
Take the annotated Seurat object as input to build a reference matrix.
}
\author{
Shuai Guo, Xiaoqian Liu, Wenyi Wang
}
