% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.wnnls.R
\name{DeMixSC.wnnls.parallel}
\alias{DeMixSC.wnnls.parallel}
\title{Parallel execution of DeMixSC.wnnls for estimating cell type proportions}
\usage{
DeMixSC.wnnls.parallel(
  bulk.mat,
  ref,
  c = 2,
  iter.max = 2000,
  eps = 0.001,
  nthread = (detectCores()[1] - 1),
  print.sample = 1,
  output.file = "./DeMixSC.save.iter.txt"
)
}
\arguments{
\item{bulk.mat}{A gene expression matrix to be deconvolved, where rows are genes and columns are samples.}

\item{ref}{A cell-type-specific gene expression matrix, where rows are genes and columns are cell types.}

\item{c}{A numeric value representing the tuning parameter for the weight function in the weighted non-negative least squares algorithm. Default is 2.}

\item{iter.max}{An integer specifying the maximum number of iterations for the deconvolution algorithm.
The algorithm will stop if this number is reached, even if convergence (steady results over iterations) has not been achieved. Default is 2000.}

\item{eps}{A numeric value representing the convergence threshold for the deconvolution algorithm.
The algorithm will stop if the difference between iterations falls below this threshold. Default is 0.001.}

\item{nthread}{An integer specifying the number of threads to use for parallel processing (executing multiple operations simultaneously).
By default, it uses one less than the total number of available CPU cores.}

\item{print.sample}{An integer specifying the index of the sample whose iterative results should be printed during the deconvolution process.
This is useful for monitoring the algorithm’s progress on a specific sample. Default is 1.
Users can set `print.sample = NULL` to disable saving iteration details.}

\item{output.file}{A character string specifying the name of the file where the algorithm’s iterative results will be saved.
This file stores intermediate deconvolution results for the specified sample. Default is `"./DeMixSC.save.iter.txt"`.}
}
\value{
\item{proportion}{A matrix of cell-type proportions, with cell types on the rows and samples on the columns.}
 \item{converge}{A vector indicating the number of iterations until convergence for each sample.}
}
\description{
Executes the DeMixSC.wnnls function in parallel for estimating cell-type proportions.
}
\note{
By default, this function utilizes the maximum number of available cores minus one for parallel computation.
This is designed to maximize computational efficiency.
However, users should be cautious when running other tasks concurrently to avoid potential system slowdowns or overloading.
If resource concerns arise or if you're operating on a shared system, users might consider specifying a lower number of threads through the `nthread` parameter.
}
