% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.wnnls.R
\name{DeMixSC.wnnls}
\alias{DeMixSC.wnnls}
\title{The weighted non-negative least square (wNNLS) framework of DeMixSC}
\usage{
DeMixSC.wnnls(
  j,
  MAT,
  R,
  c,
  iter.max = iter.max,
  eps = eps,
  print.sample = print.sample,
  output.file = output.file
)
}
\arguments{
\item{j}{Index used for parallel processing.}

\item{MAT}{A normalized expression matrix.}

\item{R}{A cell-type-specific gene expression matrix, where rows are genes and columns are cell types.}

\item{c}{A numeric value representing the tuning parameter for the weight function in the weighted non-negative least squares algorithm. Default is 2.}

\item{iter.max}{An integer specifying the maximum number of iterations for the deconvolution algorithm.
The algorithm will stop if this number is reached, even if convergence (steady results over iterations) has not been achieved. Default is 2000.}

\item{eps}{A numeric value representing the convergence threshold for the deconvolution algorithm.
The algorithm will stop if the difference between iterations falls below this threshold. Default is 0.001.}

\item{print.sample}{An integer specifying the index of the sample whose iterative results should be printed during the deconvolution process.
This is useful for monitoring the algorithm’s progress on a specific sample. Default is 1.
Users can set `print.sample = NULL` to disable saving iteration details.}

\item{output.file}{A character string specifying the name of the file where the algorithm’s iterative results will be saved.
This file stores intermediate deconvolution results for the specified sample. Default is `"./DeMixSC.save.iter.txt"`.}
}
\value{
\item{proportion}{A cell-type proportion vector for all cell types in the sample.}
 \item{converge}{Number of iterations to reach convergence or a message indicating maximum iterations reached.}
}
\description{
This function estimates the cell type proportions for each bulk sample.
The function performs the DeMixSC algorithm by iteratively estimating cell-type proportions and updating weights.
}
\author{
Shuai Guo, Xiaoqian Liu, Wenyi Wang
}
