% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_benchmark_bulk}
\alias{hgsc_benchmark_bulk}
\title{HGSC Benchmark Bulk Data}
\format{
A expression matrix with 25,370 rows and 7 columns. Rows represent genes, and columns represent samples.
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-03077-7}
}
\usage{
hgsc_benchmark_bulk
}
\description{
HGSC Benchmark Bulk Data
}
\keyword{datasets}
