% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_benchmark_ground_truth}
\alias{hgsc_benchmark_ground_truth}
\title{True Cell Type Proportions for HGSC Benchmark Data}
\format{
A matrix with 13 rows and 7 columns. Rows represent cell-types, and columns represent samples.
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-03077-7}
}
\usage{
hgsc_benchmark_ground_truth
}
\description{
True Cell Type Proportions for HGSC Benchmark Data
}
\keyword{datasets}
