% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_benchmark_pseudobulk}
\alias{hgsc_benchmark_pseudobulk}
\title{HGSC Benchmark Pseudo-bulk Data}
\format{
A expression matrix with 17,109 rows and 7 columns. Rows represent genes, and columns represent samples.
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-03077-7}
}
\usage{
hgsc_benchmark_pseudobulk
}
\description{
HGSC Benchmark Pseudo-bulk Data
}
\keyword{datasets}
