% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_benchmark_reference}
\alias{hgsc_benchmark_reference}
\title{HGSC Benchmark Reference Data}
\format{
A list with the following components:
\describe{
  \item{reference}{A numeric matrix with 29,131 rows (genes) and 13 columns (cell types). Each entry represents the expression level of a gene in a cell type.}
  \item{cell.size}{A named numeric vector with 13 elements representing the cell sizes for the 13 cell types.}
  \item{theta}{A numeric matrix with 29,131 rows (genes) and 13 columns (cell types) representing the relative abundance (theta) of each gene in each cell type.}
}
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-03077-7}
}
\usage{
hgsc_benchmark_reference
}
\description{
HGSC Benchmark Reference Data
}
\keyword{datasets}
