% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_consensus_reference}
\alias{hgsc_consensus_reference}
\title{HGSC Consensus Reference Data}
\format{
A matrix with 16,084 rows and 13 columns. Rows represent genes, and columns represent cell types.
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-023-03077-7}
}
\usage{
hgsc_consensus_reference
}
\description{
HGSC Consensus Reference Data
}
\keyword{datasets}
