% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_lee_cohort_clinical_info}
\alias{hgsc_lee_cohort_clinical_info}
\title{HGSC Lee Cohort Clinical Information}
\format{
A data frame with 41 rows and 4 variables:
\describe{
  \item{external_ID}{A character vector representing external identifiers.}
  \item{sequencing_ID}{A character vector representing sequencing experiment identifiers.}
  \item{clinical_Sample_ID}{A character vector representing the clinical sample identifiers.}
  \item{response}{A character vector representing the clinical response to neoadjuvant chemotherapy.}
}
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(20)30392-2}
}
\usage{
hgsc_lee_cohort_clinical_info
}
\description{
HGSC Lee Cohort Clinical Information
}
\keyword{datasets}
