% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_lee_cohort_mc_staining}
\alias{hgsc_lee_cohort_mc_staining}
\title{HGSC Lee Cohort Macrophage (MC) Staining Data}
\format{
A data frame with 21 rows and 3 variables:
\describe{
  \item{group}{A character vector indicating the response group.}
  \item{external_ID}{A character vector representing the external identifiers for each sample.}
  \item{Macrophages}{A numeric vector representing the quantitative measure of macrophages staining results.}
}
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(20)30392-2}
}
\usage{
hgsc_lee_cohort_mc_staining
}
\description{
HGSC Lee Cohort Macrophage (MC) Staining Data
}
\keyword{datasets}
