% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{hgsc_lee_cohort_raw_counts}
\alias{hgsc_lee_cohort_raw_counts}
\title{HGSC Lee Cohort Raw Counts}
\format{
A matrix with 17,109 rows and 41 columns. Rows represent genes, and columns represent samples.
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(20)30392-2}
}
\usage{
hgsc_lee_cohort_raw_counts
}
\description{
HGSC Lee Cohort Raw Counts
}
\keyword{datasets}
