% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_amd_cohort_clinical_info}
\alias{retina_amd_cohort_clinical_info}
\title{Retina Ratnapriya et al. AMD Cohort Clinical Information}
\format{
A data frame with 453 rows and 3 variables:
\describe{
  \item{sample_id}{A character vector representing unique sample identifiers.}
  \item{age}{A numeric vector representing the age of the individuals in the cohort.}
  \item{mgs_level}{A numeric vector representing the severity of macular degeneration, using the MGS (Macular Grading Scale) level.}
}
}
\source{
\url{https://www.nature.com/articles/s41588-019-0351-9}
}
\usage{
retina_amd_cohort_clinical_info
}
\description{
Retina Ratnapriya et al. AMD Cohort Clinical Information
}
\keyword{datasets}
