% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_amd_cohort_raw_counts}
\alias{retina_amd_cohort_raw_counts}
\title{Retina Ratnapriya et al. AMD Cohort Raw Counts}
\format{
A matrix with 55881 rows and 453 columns. Rows represent genes, and columns represent samples.
}
\source{
\url{https://www.nature.com/articles/s41588-019-0351-9}
}
\usage{
retina_amd_cohort_raw_counts
}
\description{
Retina Ratnapriya et al. AMD Cohort Raw Counts
}
\keyword{datasets}
