% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_benchmark_bulk_batch1}
\alias{retina_benchmark_bulk_batch1}
\title{Retina Benchmark Bulk Data - Batch 1}
\format{
A expression matrix with 9,558 rows and 4 columns. Rows represent genes, and columns represent samples.
}
\source{
Generated by this study.
}
\usage{
retina_benchmark_bulk_batch1
}
\description{
Retina Benchmark Bulk Data - Batch 1
}
\keyword{datasets}
