% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_benchmark_ground_truth}
\alias{retina_benchmark_ground_truth}
\title{True Cell Type Proportions for Retina Benchmark Data}
\format{
A matrix with 7 rows and 24 columns. Rows represent cell-types, and columns represent samples.
}
\source{
Generated by this study.
}
\usage{
retina_benchmark_ground_truth
}
\description{
True Cell Type Proportions for Retina Benchmark Data
}
\keyword{datasets}
