% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_benchmark_pseudobulk_batch2}
\alias{retina_benchmark_pseudobulk_batch2}
\title{Retina Benchmark Pseudo-bulk Data - Batch 2}
\format{
A expression matrix with 10,013 rows and 20 columns. Rows represent genes, and columns represent samples.
}
\source{
Generated by this study.
}
\usage{
retina_benchmark_pseudobulk_batch2
}
\description{
Retina Benchmark Pseudo-bulk Data - Batch 2
}
\keyword{datasets}
