% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_benchmark_reference_batch2}
\alias{retina_benchmark_reference_batch2}
\title{Retina Benchmark Reference Data - Batch 2}
\format{
A list with the following components:
\describe{
  \item{reference}{A numeric matrix with 12,606 rows (genes) and 7 columns (cell types). Each entry represents the expression level of a gene in a cell type.}
  \item{cell.size}{A named numeric vector with 7 elements representing the cell sizes for the 7 cell types.}
  \item{theta}{A numeric matrix with 12,606 rows (genes) and 7 columns (cell types) representing the relative abundance (theta) of each gene in each cell type.}
}
}
\source{
Generated by this study.
}
\usage{
retina_benchmark_reference_batch2
}
\description{
Retina Benchmark Reference Data - Batch 2
}
\keyword{datasets}
