% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixSC.data.document.R
\docType{data}
\name{retina_consensus_reference}
\alias{retina_consensus_reference}
\title{Retina Consensus Reference Data}
\format{
A matrix with 9993 rows and 10 columns. Rows represent genes, and columns represent cell types.
}
\source{
Generated by this study.
}
\usage{
retina_consensus_reference
}
\description{
Retina Consensus Reference Data
}
\keyword{datasets}
